[reorganization]:Move table rows logic to its own app
Fixes #7364
