+++ specifyweb/backend/table_rows/urls.py
@@ -1,0 +1,7 @@
+from django.urls import re_path
+
+from . import views
+
+urlpatterns = [ 
+   re_path(r'^specify_rows/(?P<model>\w+)/$', views.rows), # permissions added  
+]
