+++ specifyweb/specify/api.py
@@ -30,7 +30,6 @@ from django.db.models.fields import DateTimeField, FloatField, DecimalField
 from sqlalchemy.engine import Row
 
 from specifyweb.backend.permissions.permissions import (
-    enforce,
     check_table_permissions,
     check_field_permissions,
     table_permissions_checker,
@@ -1159,53 +1158,6 @@ def uri_for_model(model, id=None) -> str:
     if id is not None:
         uri += '%d/' % int(id)
     return uri
-
-class RowsForm(GetCollectionForm):
-    fields = forms.CharField(required=True) # type: ignore
-    distinct = forms.CharField(required=False)
-    defaults = dict(
-        domainfilter=None,
-        limit=0,
-        offset=0,
-        orderby=None,
-        distinct=False,
-        fields=None,
-        filterchronostrat=False,
-    )
-
-def rows(request, model_name: str) -> HttpResponse:
-    enforce(request.specify_collection, request.specify_user_agent, [f'/table/{model_name.lower()}'], "read")
-
-    form = RowsForm(request.GET)
-
-    if not form.is_valid():
-        return HttpResponseBadRequest(toJson(form.errors), content_type='application/json')
-
-    query = apply_filters(request.specify_collection, request.GET, model_name, form.cleaned_data)
-    fields = form.cleaned_data['fields'].split(',')
-    try:
-        query = query.values_list(*fields).order_by(*fields)
-    except FieldError as e:
-        return HttpResponseBadRequest(e)
-    if form.cleaned_data['domainfilter'] == 'true':
-        query = filter_by_collection(query, request.specify_collection)
-    if form.cleaned_data['orderby']:
-        try:
-            query = query.order_by(form.cleaned_data['orderby'])
-        except FieldError as e:
-            raise OrderByError(e)
-    if form.cleaned_data['distinct']:
-        query = query.distinct()
-
-    limit = form.cleaned_data['limit']
-    offset = form.cleaned_data['offset']
-    if limit == 0:
-        query = query[offset:]
-    else:
-        query = query[offset:offset + limit]
-
-    data = list(query)
-    return HttpResponse(toJson(data), content_type='application/json')
 
 def _handle_special_update_posts(obj):
     from specifyweb.backend.interactions.cog_preps import enforce_interaction_sibling_prep_max_count
