+++ specifyweb/specify/tests/test_rows.py
@@ -15,7 +15,7 @@ class TestRows(GeoTimeTestsContext):
 
     def test_simple_fetch(self):
         
-        result_rows = self.c.get("/api/specify_rows/collectionobject/", {
+        result_rows = self.c.get("/table_rows/specify_rows/collectionobject/", {
             'fields': 'catalognumber,collection__collectionname'
         })
         
@@ -46,7 +46,7 @@ class TestRows(GeoTimeTestsContext):
             self.collectionobjects[4],
             dict(text1="TextData2")
         )
-        result_rows = self.c.get("/api/specify_rows/collectionobject/", {
+        result_rows = self.c.get("/table_rows/specify_rows/collectionobject/", {
             'fields': 'catalognumber,text1',
             'orderby': 'text1'
         })
@@ -87,7 +87,7 @@ class TestRows(GeoTimeTestsContext):
             raise error
 
 
-        distinct_result_rows = self.c.get("/api/specify_rows/collectionobject/", {
+        distinct_result_rows = self.c.get("/table_rows/specify_rows/collectionobject/", {
             'fields': 'text1',
             'orderby': 'text1',
             'distinct': True
@@ -98,7 +98,7 @@ class TestRows(GeoTimeTestsContext):
 
     def test_limit(self):
 
-        result_rows = self.c.get("/api/specify_rows/collectionobject/", {
+        result_rows = self.c.get("/table_rows/specify_rows/collectionobject/", {
             'fields': 'catalognumber,collection__collectionname',
             'limit': 3
         })
@@ -110,7 +110,7 @@ class TestRows(GeoTimeTestsContext):
 
     def test_offset(self):
 
-        result_rows = self.c.get("/api/specify_rows/collectionobject/", {
+        result_rows = self.c.get("/table_rows/specify_rows/collectionobject/", {
             'fields': 'catalognumber,collection__collectionname',
             'limit': 3,
             'offset': 1
@@ -137,7 +137,7 @@ class TestRows(GeoTimeTestsContext):
             ) for i in range(5)
         ]
 
-        result_rows = self.c.get("/api/specify_rows/collectionobject/", {
+        result_rows = self.c.get("/table_rows/specify_rows/collectionobject/", {
             'fields': 'catalognumber,collection__collectionname',
         })
 
@@ -149,7 +149,7 @@ class TestRows(GeoTimeTestsContext):
             [*result_1, *result_2]
         )
 
-        domain_result_rows = self.c.get("/api/specify_rows/collectionobject/", {
+        domain_result_rows = self.c.get("/table_rows/specify_rows/collectionobject/", {
             'fields': 'catalognumber,collection__collectionname',
             'domainfilter': 'true'
         })
@@ -178,7 +178,7 @@ class TestRows(GeoTimeTestsContext):
         )
 
 
-        result_rows = self.c.get("/api/specify_rows/collectionobject/", {
+        result_rows = self.c.get("/table_rows/specify_rows/collectionobject/", {
             'fields': 'preparations__remarks,catalognumber',
             'domainfilter': 'true',
             'preparations__remarks__isnull': 'false'
@@ -189,7 +189,7 @@ class TestRows(GeoTimeTestsContext):
             [co_2_prep_2.remarks, self.collectionobjects[1].catalognumber]
         ])
 
-        result_rows = self.c.get("/api/specify_rows/collectionobject/", {
+        result_rows = self.c.get("/table_rows/specify_rows/collectionobject/", {
             'fields': 'preparations__remarks,catalognumber',
             'domainfilter': 'true',
             'preparations__remarks__isnull': 'true'
@@ -229,7 +229,7 @@ class TestRows(GeoTimeTestsContext):
         self._numeric_setup()
 
 
-        result_rows = self.c.get("/api/specify_rows/collectionobject/", {
+        result_rows = self.c.get("/table_rows/specify_rows/collectionobject/", {
             'fields': 'integer1,catalognumber',
             'integer1__in': '3,8'
         })
@@ -247,7 +247,7 @@ class TestRows(GeoTimeTestsContext):
             _get_row(3)
         ])
 
-        result_rows = self.c.get("/api/specify_rows/collectionobject/", {
+        result_rows = self.c.get("/table_rows/specify_rows/collectionobject/", {
             'fields': 'text1,catalognumber',
             'text1__in': 'NewText1,ModText4'
         })
@@ -269,7 +269,7 @@ class TestRows(GeoTimeTestsContext):
     def test_range_query(self):
         
         self._numeric_setup()
-        result_rows = self.c.get("/api/specify_rows/collectionobject/", {
+        result_rows = self.c.get("/table_rows/specify_rows/collectionobject/", {
             'fields': 'integer1,catalognumber',
             'integer1__range': '4,8'
         })
@@ -311,7 +311,7 @@ class TestRows(GeoTimeTestsContext):
             endperiod=50,
             enduncertainty=None,
         )
-        result_rows = self.c.get("/api/specify_rows/geologictimeperiod/", {
+        result_rows = self.c.get("/table_rows/specify_rows/geologictimeperiod/", {
             'fields': 'name,rankid',
             'filterchronostrat': False
         })
@@ -329,7 +329,7 @@ class TestRows(GeoTimeTestsContext):
 
         self.assertCountEqual(expected_rows, json.loads(result_rows.content.decode()))
 
-        result_rows = self.c.get("/api/specify_rows/geologictimeperiod/", {
+        result_rows = self.c.get("/table_rows/specify_rows/geologictimeperiod/", {
             'fields': 'name,rankid',
             'filterchronostrat': True
         })
