+++ specifyweb/specify/urls.py
@@ -25,7 +25,6 @@ urlpatterns = [
     re_path(r'^specify/bulk/(?P<model>\w+)/(?P<copies>\d+)/$', views.collection_bulk_copy), # permissions added
     re_path(r'^specify/bulk/(?P<model>\w+)/$', views.collection_bulk), # permissions added
     re_path(r'^specify/(?P<model>\w+)/$', views.collection), # permissions added
-    re_path(r'^specify_rows/(?P<model>\w+)/$', views.rows), # permissions added  
 
     re_path(r'^delete_blockers/(?P<model>\w+)/(?P<id>\d+)/$', views.delete_blockers),
