+++ specifyweb/urls.py
@@ -24,6 +24,7 @@ from specifyweb.backend.stats import urls as stat_urls
 from specifyweb.backend.trees import urls as trees_urls
 from specifyweb.backend.merge import urls as merge_urls
 from specifyweb.backend.locality_update_tool import urls as locality_update_tool_urls
+from specifyweb.backend.table_rows import urls as table_rows_urls
 
 # print(get_resolver().reverse_dict.keys()) # Use for debugging urls
 
@@ -69,5 +70,6 @@ urlpatterns = [
     path('stats/', include(stat_urls)),
     path('trees/', include(trees_urls)),
     path('merge/', include(merge_urls)),
-    path('locality_update_tool/', include(locality_update_tool_urls))
+    path('locality_update_tool/', include(locality_update_tool_urls)),
+    path('table_rows/', include(table_rows_urls))
 ]
